refactor: Move configuration files and adjust imports accordingly
