--- ormconfig.js
@@ -1,9 +1,0 @@
-module.exports = {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
-  ...require('./typeorm.config').typeOrmConfig,
-  entities: ['dist/**/*.entity.js'],
-  migrations: ['dist/migration/*.js'],
-  cli: {
-    migrationsDir: 'src/migration',
-  },
-};
