+++ src/app.module.ts
@@ -1,4 +1,4 @@
-import { redisConfig } from '../redis.config';
+import { redisConfig } from './config/redis.config';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { MiddlewareConsumer, Module, NestModule, OnModuleInit } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -22,7 +22,7 @@ import { SeederService } from './modules/seeder/seeder.service';
 import { JwtInterceptor } from './common/interceptros/jwt.interceptor';
 import { DeviceInterceptor } from './common/interceptros/device.interceptor';
 import { TimezoneMiddleware } from './common/middlewares/timezone.middleware';
-import { TypeOrmOptions } from '../typeorm.options';
+import { TypeormOptionsConfig } from './config/typeormOptions.config';
 import { ViewModule } from './modules/view/view.module';
 
 @Module({
@@ -31,7 +31,7 @@ import { ViewModule } from './modules/view/view.module';
       isGlobal: true,
       envFilePath: '../.env',
     }),
-    TypeOrmModule.forRootAsync(TypeOrmOptions),
+    TypeOrmModule.forRootAsync(TypeormOptionsConfig),
     RedisModule.forRootAsync({
       useFactory: () => redisConfig,
     }),
